Pre-prod v. Mistral work. Local machine avaliable to answer questions.
