+++ .env.example
@@ -27,10 +27,9 @@ ADMIN_PANEL_HOST=adminpanel:80
 VK_ACCESS_GROUP_TOKEN=
 TG_ACCESS_TOKEN=
 
-# настройки API Sber GigaChat
-GIGACHAT_MODEL=GigaChat-Pro
-GIGACHAT_TOKEN=
-GIGACHAT_SCOPE=
+# настройки API Mistral
+MISTRAL_API=
+MISTRAL_MODEL=
 
 # параметры подключения к Confluence
 CONFLUENCE_TOKEN=
