+++ docs/README.md
@@ -64,7 +64,7 @@
 
 Архитектура приложения (рис. 1) состоит из четырёх составляющих: [микросервис работы с LongPoll API чат-ботов VK и Telegram (Chatbot)](chatbot.md), [микросервис ответа на вопрос по документам (QA)](qa.md), [веб-приложение, с которым взаимодействует администратор (Adminpanel)](adminpanel.md), база данных. Каждая составляющая вынесена в отдельный Docker-контейнер.
 
-Chatbot взаимодействует с API [вики-системы Confluence](confluence-integration.md) для получения справочной информации и LongPoll API VK и Telegram. QA взаимодействует с API вики-системы Confluence для получения документов при формировании векторного хранилища документов и API большой языковой модели сервиса GigaChat от Сбера для генерации ответов. Панель администратора взаимодействует с микросервисами и базой данных, чтобы провести аналитику вопросов. Также администратор может произвести ручную синхронизацию документов на Confluence.
+Chatbot взаимодействует с API [вики-системы Confluence](confluence-integration.md) для получения справочной информации и LongPoll API VK и Telegram. QA взаимодействует с API вики-системы Confluence для получения документов при формировании векторного хранилища документов и API большой языковой модели Mistral для генерации ответов. Панель администратора взаимодействует с микросервисами и базой данных, чтобы провести аналитику вопросов. Также администратор может произвести ручную синхронизацию документов на Confluence.
 
 > [!IMPORTANT]
 > В отдельный Docker-контейнер `db-migrate` вынесен сервис создания миграций баз данных. Синхронизация между базой данных и моделями ORM микросервисов должна осуществляться путём [миграций](https://alembic.sqlalchemy.org/en/latest/ops.html).
